Split background and video to separate objects
